fix icon moving bug
if user attempts to move the mouse very quickly, it leaves the element but the element never gets the mouse released check, leaving the icon in a buggy state. it gets 'sticky', so when the mouse cursor moves back into the frame, because it never stopped tracking, it then starts tracking and following the mouse cursor despite the mouse being down. a smarter check was done where we check the entire document for mouse checks instead, so that regardless of mouse speed or unoptimisation from either the code or the browser would not affect icon movement.